chore: flatten utils/helper directories - esbuild-plugin-file-path-extensions can't handle them
