+++ packages/all/package.json
@@ -41,7 +41,6 @@
     "email"
   ],
   "peerDependencies": {
-    "@jsx-email/cli": "^1.0.0",
     "react": "^18.2.0"
   },
   "dependencies": {
