+++ packages/body/package.json
@@ -29,8 +29,6 @@
     "dist/**"
   ],
   "peerDependencies": {
-    "@jsx-email/all": "^2.0.0",
-    "@jsx-email/cli": "^1.0.0",
     "react": "^18.2.0"
   },
   "devDependencies": {
