--- packages/button/src/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './px-to-pt';
-export * from './parse-padding';
