--- packages/button/src/utils/px-to-pt.ts
@@ -1,2 +1,0 @@
-export const pxToPt = (px: number): number | null =>
-  typeof px === 'number' && !isNaN(Number(px)) ? (px * 3) / 4 : null;
