--- packages/heading/src/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './as';
-export * from './spaces';
