+++ packages/markdown/package.json
@@ -39,9 +39,6 @@
     "markdown"
   ],
   "peerDependencies": {
-    "@jsx-email/all": "^2.0.0",
-    "@jsx-email/cli": "^1.0.0",
-    "@jsx-email/render": "^1.0.0",
     "react": "^18.2.0"
   },
   "dependencies": {
