+++ packages/render/src/escape-string.ts
