+++ packages/render/src/index.ts
@@ -1,2 +1,2 @@
 export * from './render';
-export * from './helpers/jsx-to-string';
+export * from './jsx-to-string';
