+++ packages/tailwind/package.json
@@ -39,8 +39,6 @@
     "tailwind"
   ],
   "peerDependencies": {
-    "@jsx-email/all": "^2.0.0",
-    "@jsx-email/cli": "^1.0.0",
     "react": "^18.2.0"
   },
   "dependencies": {
